+++ __tests__/components/Subscribe.test.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
-import Label from './Label';
+import Subscribe from './../../app/Components/Steps/Subscribe';
 import renderer from 'react-test-renderer';
-import {shallow} from 'enzyme';
 
 test('Take Snapshot', () => {
   const component = renderer.create(
-    <Label
-      activeStep={2}
-      repo="kbariotis/ansible-nodejs-digitalocean"
+    <Subscribe
+      activeStep={5}
+      repo="https://github.com/kbariotis/ansible-nodejs-digitalocean"
+      label="enhancement"
+      email="konmpar@gmail.com"
       next={function(){}}
       startOver={function(){}}
       styles={{}}/>
